envers의 revision 조회 시, 이력 번호가 20억이 넘을 경우의 대응.

+++ src/main/java/me/gogosing/persistence/repository/PlaylistInventoryRepository.java
@@ -14,5 +14,5 @@ import org.springframework.stereotype.Repository;
 public interface PlaylistInventoryRepository extends
         JpaRepository<PlaylistInventoryEntity, Long>,
         PlaylistInventoryRepositoryCustom,
-        RevisionRepository<PlaylistInventoryEntity, Long, Integer> {
+        RevisionRepository<PlaylistInventoryEntity, Long, Long> {
 }
